updated totalAmount type in PreCheckoutQuery from 64bit integer to 32bit.

+++ include/tgbotxx/objects/PreCheckoutQuery.hpp
@@ -24,7 +24,7 @@ namespace tgbotxx {
         /// @brief Total price in the smallest units of the currency (integer, not float/double).
         /// For example, for a price of US$ 1.45 pass amount = 145. See the exp parameter in currencies.json,
         /// it shows the number of digits past the decimal point for each currency (2 for the majority of currencies).
-        std::int64_t totalAmount;
+        std::int32_t totalAmount;
 
         /// @brief Bot specified invoice payload
         std::string invoicePayload;
